requirements.txt
pandas~=1.5.2
SQLAlchemy~=1.4.45
xlwings~=0.28.8
PyYAML~=6.0
colorlog~=6.7.0
urllib3~=1.26.14
exchangelib~=4.9.0
requests~=2.28.2
APScheduler~=3.10.0
